Simplified types

+++ clr/color.go
@@ -3,16 +3,16 @@ package clr
 // Color is the primary interface this library defines. Hex, HSL and RGB Types all satisfy this interface.
 type Color interface {
 	Valid() bool
-	RGB() (uint8, uint8, uint8)
-	HSL() (uint16, uint8, uint8)
-	HSV() (uint16, uint8, uint8)
-	CMYK() (uint8, uint8, uint8, uint8)
+	RGB() (int, int, int)
+	HSL() (int, int, int)
+	HSV() (int, int, int)
+	CMYK() (int, int, int, int)
 	XYZ() (float64, float64, float64)
 	CIELAB() (l, a, b float64)
 	Hex() string
 	ColorName(colors ColorTable) ColorSpace
 	Distance(c Color) float64
-	RGBA() (r, g, b, a uint8)
+	RGBA() (r, g, b, a int)
 }
 
 // ColorTable allows colors to be mapped to names, this must be fulfilled to the library as there is no built in ColorTable.
