+++ clr/hsv_test.go
@@ -23,7 +23,7 @@ func colorEquality(a RGB, b RGB) bool {
 	return abs(a.R, b.R) <= 1 && abs(a.G, b.G) <= 1 && abs(a.B, b.B) <= 1
 }
 
-func abs(a uint8, b uint8) int {
+func abs(a int, b int) int {
 	c := int(a) - int(b)
 	if c < 0 {
 		c *= -1
